clean third code

+++ src/utils/common.h
@@ -22,8 +22,8 @@ typedef short           int16_t;
 typedef unsigned short  uint16_t;
 typedef int             int32_t;
 typedef unsigned int    uint32_t;
-typedef long            int64_t;
-typedef unsigned long   uint64_t;
+typedef long long       int64_t;
+typedef unsigned long long   uint64_t;
 
 // for repeat patterns
 #define SS_REPEAT1(X) SS_REPEAT_PATTERN(X)
